Delegate Collection Property Management to CollectionPropertiesZkStateReader
